pyproject.toml
[project]
name = "targetpointer"
version = "0.4.0"
description = "Fixed-camera person pointing demo with YOLO, serial control, Blue Pill, and SG90"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[elevenlabs,openai]~=1.4",
    "openai>=2.0.0",
    "opencv-python>=4.13.0.92",
    "platformio>=6.1.19",
    "python-dotenv>=1.2.1",
    "pyside6>=6.9.1",
    "pyserial>=3.5",
    "reportlab>=4.4.5",
    "ultralytics>=8.4.37",
    "elevenlabs>=2.44.0",
]
